+++ pos_restaurant_web_menu/doc/RELEASE_NOTES.md
@@ -4,3 +4,11 @@
 #### Version 16.0.1.0.0
 #### ADD
 - Initial commit for POS Restaurant Web Menu
+
+#### 23.08.2024
+#### Version 16.0.1.0.1
+#### UPDATE
+- Update the order menu screen scroll option
+- Update the feature for delete the order line
+
+